LIBFCREPO-1073. Updated docs to use full URL for stub "--access" flag
Modified "Add_Plant_Patents_PDF_Scans.md" to use the full URL
"http://vocab.lib.umd.edu/access#Public" for the "--access" flag
when running the "stub" command, instead of using "umdaccess:Public".

This appears to be necessary, because the "umdaccess" prefic to the
full URL mapping was not being made (likely because the "get_manager"
method in the "plastron/namespaces/__init__.py" file is not being
called by the Plastron "stub" command).

https://issues.umd.edu/browse/LIBFCREPO-1073