LIBITD-2097. Added working file to .gitignore, and tweaked procedure
Added the "data_updated.csv" file generated by the
"Patents Metadata Update" procedure to ".gitignore", so that it would
not get accidentally checked in.

Also tweaked the "Patents Metadata Update" procedure to fix a step
that was copy/pasted from the "Add Plant Patents PDF Scans",
changing it to delete the working file created by the
"Patents Metadata Update" procedure.

https://issues.umd.edu/browse/LIBITD-2097