LIBITD-2426. Updated README.md for M-Series (Apple Silicon) MacBooks
Updated the Docker build instructions in the README.md to use Kubernetes
to build the Docker image, which is what should always be used for
production use.

Added information about using the “SOLR_JAVA_STACK_SIZE” environment
variable if the Kubernetes-built Docker image does not run successfully
on an M-Series MacBook.

https://umd-dit.atlassian.net/browse/LIBITD-2426